feat(multi-client): Split the Editor data structure to support multiple clients
This commit is the first step towards allowing the editor to support
multiple clients, where a "client" can be thought of as a window
displaying the editor UI. This change may support the following use cases:

- Client-server architecture with multiple terminal windows displaying the
  UI with shared buffers (#312). This will be the initial use case.
- Helix mode for other editors.
- Multiple windows for a GUI.

With this change, the Editor owns multiple EditorClients, with each
EditorClient holding the data for a particular client, such as its
views and state related to user input, such as the mode and repeat
count. Clients are referenced using a ClientId which is used as a key
for per-client data.

For the time being, the Application class only creates one client, but
in a followup change (which is still work-in-progress and is available
at [1]) the Application will be modified to implement the client-server
support.

Apologies for the size of this change, I couldn't see a good way to split
it up due to the many interdependencies. Probably the easiest way to
review it is to download it locally and use `git show -w --color-words`.

Note that only a minimum amount of per-client state has been moved to
the EditorClient; for example, the state for the status bar is still
shared. The remaining state will be moved in followup changes.

[1] https://github.com/pcc/helix/tree/client-server-rebase