feat(multi-client): Further improve :q behavior
This is another customization from my Kakoune fork that I can't do
without. The issue is when running `git commit` where $EDITOR is
a multi-client Helix. Because git will overwrite the COMMIT_EDITMSG
file for each new commit, Helix will think that the buffer is externally
modified (which is true) if it was edited in a previous client and refuse
to write it by default. The fix is as follows: if the file being closed
with :q is unmodified and there are no other views currently showing it,
we forget the buffer. It is unlikely that it will be useful to remember
the buffer, at least not if your workflow resembles mine. It might be
worth adding a customization for this for people with different workflows.